Initial work on replacing sassc with jsass
This commit introduces the jsass libsass wrapper as the backend for compiling
SASS/SCSS stylesheets.

The benefits of using jsass instead of sassc is that the user will not have to
download and install sassc before using this module.

The jsass library already includes compiled versions of libsass for Windows,
Linux and macOS.

The other change that was made was to bring the :line-numbers option name inline
with libsass's terminology by renaming it to :source-comments.

This option includes comments in the outputted .css file that indicate the file
and line number that generated the following CSS code (sort of a poor-man's
source map)