Make the output-uri file in the same folder as the entrypoint
Because of how libsass generates sourcemaps, when passing the input and output
file URI's to the compile method, we want both the input and output files to be
in the same absolute directory even though the output .css file will actually be
written in a different directory.

This causes the source map to be generated correctly.